refactor(materialization): split hypertable write ownership
